backend/requirements.txt
fastapi
uvicorn[standard]
torch
transformers
newsapi-python
numpy
scikit-learn
joblib
requests
